+++ model/src/main/java/io/streamzi/openshift/dataflow/crds/ProcessorList.java
@@ -1,4 +1,4 @@
-package io.streamzi.openshift.dataflow.model.crds;
+package io.streamzi.openshift.dataflow.crds;
 
 import io.fabric8.kubernetes.client.CustomResourceList;
