+++ watcher/src/main/java/io/streamzi/openshift/FlowController.java
@@ -6,9 +6,9 @@ import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.fabric8.kubernetes.api.model.apiextensions.CustomResourceDefinition;
 import io.fabric8.openshift.api.model.DeploymentConfig;
 import io.streamzi.openshift.dataflow.model.ProcessorFlow;
-import io.streamzi.openshift.dataflow.model.crds.DoneableFlow;
-import io.streamzi.openshift.dataflow.model.crds.Flow;
-import io.streamzi.openshift.dataflow.model.crds.FlowList;
+import io.streamzi.openshift.dataflow.crds.DoneableFlow;
+import io.streamzi.openshift.dataflow.crds.Flow;
+import io.streamzi.openshift.dataflow.crds.FlowList;
 
 import java.util.HashMap;
 import java.util.List;
